% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xc_summary.R
\name{xc_get_last_pg_num}
\alias{xc_get_last_pg_num}
\title{Scrape x value of last page number}
\usage{
xc_get_last_pg_num(page)
}
\arguments{
\item{page}{xml object resulting from html_read}
}
\value{
numeric scalar for last x value needed
}
\description{
Scrape x value of last page number
}
\examples{
url_pg <- "https://www.xcontest.org/world/en/pilots/"
page <- xml2::read_html(url(url_pg))
last_num <- xc_get_last_pg_num(page)
}
