% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xc_summary.R
\name{xc_get_pilot_tbl}
\alias{xc_get_pilot_tbl}
\title{Get pilot table from website}
\usage{
xc_get_pilot_tbl(page)
}
\arguments{
\item{page}{xml object resulting from html_read}
}
\value{
tidy dataframe containing num, full_name, nickname, country, num_flights,
open_score, fail_score, registered and info variables for each pilot on the page
}
\description{
Get pilot table from website
}
\examples{
url_pg <- "https://www.xcontest.org/world/en/pilots/"
page <- xml2::read_html(url(url_pg))
df_pg1 <- xc_get_pilot_tbl(page)
}
