% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xc_summary.R
\name{xc_summary}
\alias{xc_summary}
\title{Cross-country paragliding data web scraping}
\usage{
xc_summary(limit_run = FALSE)
}
\arguments{
\item{limit_run}{either TRUE or FALSE(default) specifies whether to pull
only 3 pages if TRUE or all pages if FALSE}
}
\value{
a dataframe containing the number, full_name, nickname, country,
num_flights, open_score, fail_score, registered, and info from the website
  \url{https://www.xcontest.org/world/en/pilots/}.
}
\description{
Cross-country paragliding data web scraping
}
\examples{
pilot_df <- xc_summary()
small_pilot_df <- xc_summary(TRUE)

}
