% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xflyr.R
\docType{package}
\name{xflyr}
\alias{xflyr}
\title{xflyr package}
\description{
Get summary of info on cross country paragliding pilots.
Eventually this will get detailed info and analyze it.
}
\author{
Jane Wall \email{djwall@yahoo.com}
}
